requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
annotated-types==0.7.0
attrs==25.4.0
certifi==2026.2.25
charset-normalizer==3.4.4
frozenlist==1.8.0
googleapis-common-protos==1.72.0
grpcio==1.78.0
idna==3.11
importlib_metadata==8.7.1
multidict==6.7.1
opentelemetry-api==1.40.0
opentelemetry-sdk==1.40.0
opentelemetry-semantic-conventions==0.61b0
packaging==25.0
propcache==0.4.1
protobuf==6.33.5
pydantic==2.12.5
pydantic_core==2.41.5
python-dotenv==1.2.2
requests==2.32.5
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
xai-sdk==1.8.0
yarl==1.23.0
zipp==3.23.0
